+++ src/Credfeto.Dispatcher.Discord.DataTypes/DiscordEmbed.cs
@@ -1,0 +1,7 @@
+using System;
+using System.Diagnostics;
+
+namespace Credfeto.Dispatcher.Discord.DataTypes;
+
+[DebuggerDisplay("{Title}: {Url}")]
+public sealed record DiscordEmbed(string Title, string Description, Uri Url, int Color);