+++ src/Credfeto.Dispatcher.Discord/Configuration/DiscordOptions.cs
@@ -1,0 +1,12 @@
+using System;
+using System.Diagnostics;
+
+namespace Credfeto.Dispatcher.Discord.Configuration;
+
+[DebuggerDisplay("WebhookUrl: {WebhookUrl}")]
+public sealed class DiscordOptions
+{
+    public Uri? WebhookUrl { get; init; }
+
+    public Uri? NotificationsChannelWebhookUrl { get; init; }
+}