+++ src/Credfeto.Dispatcher.GitHub.DataTypes/NotificationRepository.cs
@@ -1,0 +1,7 @@
+using System;
+using System.Diagnostics;
+
+namespace Credfeto.Dispatcher.GitHub.DataTypes;
+
+[DebuggerDisplay("{FullName}: {Url}")]
+public sealed record NotificationRepository(string FullName, Uri Url);